Respawning + Statistics management
Players now respawn if their health drops to zero

When a player kills another player, the stats are updated

+++ Assets/Scripts/Statistics/PlayerMatchStatisticsData.cs
@@ -32,7 +32,7 @@ public struct PlayerMatchStatisticsData : INetworkSerializable, IEquatable<Playe
         }
         else
         {
-            return kills / deaths;
+            return (float) kills / deaths;
         }
     }
